Clean up where function is located
